Implemented BaseTopology, refactored AkkaTopology
